# Conflicts:
#	src/main/java/com/springbootcallingexternalapi/LeagueOfLegends/Exceptions/AccountExceptions/AccountDataException.java
#	src/main/java/com/springbootcallingexternalapi/LeagueOfLegends/RestControllers/RiotRestController.java
#	src/main/java/com/springbootcallingexternalapi/LeagueOfLegends/Services/RiotRequestorService.java
#	src/test/java/com/springbootcallingexternalapi/LeagueOfLegends/RestControllers/RiotRestControllerTest.java

+++ src/main/java/com/springbootcallingexternalapi/LeagueOfLegends/Repositories/AccountRepository.java
@@ -26,7 +26,6 @@ public class AccountRepository {
     private JdbcTemplate jdbcTemplate;
 
     public void insertAccount(AccountBaseModel account, Integer owner) throws AccountDataException, AccountExistsOrNotException {
-
         String sql = "INSERT INTO \"Account\"(id, puuid, accountid, \"revisionDate\", \"owner\", name) VALUES (?, ?, ?, ?, ?, ?);";
         Object[] params = {
                 account.getId(),
@@ -47,7 +46,6 @@ public class AccountRepository {
     public void deleteAccount( String account,Integer ownerID) throws CharacterNotAllowedException, AccountNotFoundException {
         String sql = "DELETE FROM \"Account\" WHERE Lower(name) =? and \"owner\" =?;";
         Object[] params = {account.toLowerCase(Locale.ROOT), ownerID};
-
         if (isAlpha(account)) {
             int result = jdbcTemplate.update(sql, params);
             if (result == 0) {
@@ -59,7 +57,6 @@ public class AccountRepository {
     public List<AccountModel> retrieveAccountByOwner(String owner,Integer ownerID) throws CharacterNotAllowedException, OwnerNotFoundException {
         String sql = "SELECT * FROM \"Account\" WHERE  \"owner\" =?";
         Object[] params = {ownerID};
-
         if (isAlpha(owner)) {
             List<AccountModel> listAccounts = jdbcTemplate.query(sql, params,
                     BeanPropertyRowMapper.newInstance(AccountModel.class));
@@ -101,7 +98,6 @@ public class AccountRepository {
     public List<AccountModel> retrieveAccountByAccountName(String account) throws CharacterNotAllowedException, AccountNotFoundException {
         String sql = "SELECT * FROM \"Account\" WHERE LOWER (name)=?";
         Object[] params = {account};
-
         if (isAlpha(account)) {
             List<AccountModel> listAccounts = jdbcTemplate.query(sql, params,
                     BeanPropertyRowMapper.newInstance(AccountModel.class));
@@ -111,12 +107,12 @@ public class AccountRepository {
         } else throw new CharacterNotAllowedException(account);
     }
 
-    public Long retrieveAccountIdByAccountName(String accountName) throws  CharacterNotAllowedException, AccountNotFoundException {
+    public Integer retrieveAccountIdByAccountName(String accountName) throws  CharacterNotAllowedException, AccountNotFoundException {
         String sql = "SELECT \"id_BD\" FROM \"Account\" WHERE LOWER(\"name\")=?;";
         Object[] params = {accountName.toLowerCase(Locale.ROOT)};
         if (isAlpha(accountName)) {
             try {
-                return jdbcTemplate.queryForObject(sql, params, Long.class);
+                return jdbcTemplate.queryForObject(sql, params, Integer.class);
             } catch (EmptyResultDataAccessException e) {
                 throw new AccountNotFoundException(accountName);
             }
